ref(issues): Refactor group activity
Group activity component reads projectId from group instead of params.
Makes it reusable at org level routes where project won't be in the URL.
